DFJJpro/requirements.txt
Django==1.10.6
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
Pillow==6.0.0
six==1.12.0
